refactor: add useImageExporter hooks
