+++ src/hooks/useShareDialog.js
@@ -1,4 +1,4 @@
-import { useState, useRef } from "react";
+import { useState, useCallback, useRef } from "react";
 
 import { useDisclosure } from "@chakra-ui/react";
 import {
@@ -10,7 +10,7 @@ import {
   AlertDialogCloseButton,
 } from "@chakra-ui/react";
 
-import Instagram from "@features/ShareWidget/components/Instagram";
+import Instagram from "@features/ShareWidget/Instagram";
 
 export const useShareDialog = () => {
   const { isOpen, onOpen, onClose } = useDisclosure();
@@ -24,35 +24,37 @@ export const useShareDialog = () => {
     onOpen();
   };
 
-  const ShareDialog = ({ content }) => {
-    return (
-      <>
-        <AlertDialog
-          motionPreset="slideInBottom"
-          leastDestructiveRef={cancelRef}
-          onClose={closeShareDialog}
-          isOpen={isOpen}
-          isCentered
-          size="md"
-        >
-          <AlertDialogOverlay />
-          <AlertDialogContent borderRadius={25} marginX={4}>
-            <AlertDialogHeader>Share to your social media</AlertDialogHeader>
-            <AlertDialogCloseButton />
-            <AlertDialogBody
-              display={`flex`}
-              flexDirection={`column`}
-              justifyContent={`center`}
-              alignItems={`center`}
-            >
-              <Instagram content={content} />
-            </AlertDialogBody>
-          </AlertDialogContent>
-        </AlertDialog>
-      </>
-    );
-  };
-  // [isOpen]
+  const ShareDialog = useCallback(
+    ({ content }) => {
+      return (
+        <>
+          <AlertDialog
+            motionPreset="slideInBottom"
+            leastDestructiveRef={cancelRef}
+            onClose={closeShareDialog}
+            isOpen={isOpen}
+            isCentered
+            size="md"
+          >
+            <AlertDialogOverlay />
+            <AlertDialogContent borderRadius={25} marginX={4}>
+              <AlertDialogHeader>Share to your social media</AlertDialogHeader>
+              <AlertDialogCloseButton />
+              <AlertDialogBody
+                display={`flex`}
+                flexDirection={`column`}
+                justifyContent={`center`}
+                alignItems={`center`}
+              >
+                <Instagram content={content} />
+              </AlertDialogBody>
+            </AlertDialogContent>
+          </AlertDialog>
+        </>
+      );
+    },
+    [isOpen]
+  );
 
   return { openShareDialog, ShareDialog };
 };
